python: clean __init__.py
