+++ bindings/python/pinocchio/__init__.py
@@ -18,59 +18,12 @@ try:
     from .pinocchio_pywrap import *
     from .pinocchio_pywrap import __version__, __raw_version__
 except ImportError:
-    import os
-    import sys
     import platform
-    import contextlib
     if platform.system() == "Windows":
-        __pinocchio_paths = os.getenv('PINOCCHIO_WINDOWS_DLL_PATH')
-        if __pinocchio_paths is None:
-            # Standard site-packages to bin path
-            RELATIVE_DLL_PATH = "..\\..\\..\\bin"
-            __dll_paths = [os.path.join(os.path.dirname(__file__), RELATIVE_DLL_PATH)]
-        else:
-            __dll_paths = __pinocchio_paths.split(os.pathsep)
+        from .windows_dll_manager import get_dll_paths, build_directory_manager
 
-        # Restore PATH state after importing Pinocchio
-        class __PathManager(contextlib.AbstractContextManager):
-            def add_dll_directory(self, dll_dir: str):
-                os.environ["PATH"] += os.pathsep + dll_dir
-
-            def __enter__(self):
-                self.old_path = os.environ["PATH"]
-                return self
-
-            def __exit__(self):
-                os.environ["PATH"] = self.old_path
-
-        # Restore DllDirectory state after importing Pinocchio
-        class __DllDirectoryManager(contextlib.AbstractContextManager):
-            def add_dll_directory(self, dll_dir: str):
-                # add_dll_directory can fail on relative path and non
-                # existing path.
-                # Since we don't know all the fail criterion we just ignore
-                # thrown exception
-                try:
-                    self.dll_dirs.append(os.add_dll_directory(dll_dir))
-                except OSError:
-                    pass
-
-            def __enter__(self):
-                self.dll_dirs = []
-                return self
-
-            def __exit__(self, *exc_details):
-                for d in self.dll_dirs:
-                    d.close()
-
-        def __build_directory_manager():
-            if sys.version_info >= (3, 8):
-                return __DllDirectoryManager()
-            else:
-                return __PathManager()
-
-        with __build_directory_manager() as dll_dir_manager:
-            for p in __dll_paths:
+        with build_directory_manager() as dll_dir_manager:
+            for p in get_dll_paths():
                 dll_dir_manager.add_dll_directory(p)
             from .pinocchio_pywrap import *
             from .pinocchio_pywrap import __version__, __raw_version__
